reorganize trtl tests
